dziala jeszcze lepiej

+++ proc2.c
@@ -14,7 +14,7 @@ void handle_signal();
 
 
 int main(int argc, char* argv[]) {
-    set_signals(signal_handler);
+    set_signals_p2(signal_handler);
 
     if (argc != 3) return 2134;
     p = atoi(argv[1]);
@@ -35,66 +35,71 @@ int main(int argc, char* argv[]) {
 }
 
 void loop(int sem_id, struct shared* shm, int msg_id) {
-
     struct msgbuf msg;
     while (1) {
         if (got_signal) handle_signal();
 
         if (is_paused) {
+            pause();
             continue;
         }
 
-        char buffer[BUFFER_SIZE];
         P_FULL;
         P_MUTEX;
+        /*
+        if (P_FULL == -1) {
+            if (errno == EINTR) continue;
+            perror("P_FULL"); break;
+        }
+        if (P_MUTEX == -1) {
+            if (errno == EINTR) { V_FULL; continue; }
+            perror("P_MUTEX"); break;
+        }
+        */
+
+            char buffer[BUFFER_SIZE];
             strcpy(buffer, shm->buf);
-            if(buffer[0] == 0) break;
+            int is_eof = (buffer[0] == 0);
         V_MUTEX;
         V_EMPTY;
 
-        
+        if (is_eof) break;
+
         msg.type = 1;
         msg.data = strlen(buffer);
         msgsnd(msg_id, &msg, sizeof(msg.data), 0);
     }
+    
+    // Sygnał końca danych do P1
     msg.type = 1;
     msg.data = -1;
     msgsnd(msg_id, &msg, sizeof(msg.data), 0);
-
 }
 
 void signal_handler(int sig) {
-    got_signal = 1;
     s = sig;
+    got_signal = 1;
 }
 
 void handle_signal() {
-    switch(s) {
-        case SIGUSR1:
-            kill(p1_id, SIGUSR1);
-            char buf[4];
-            int signal_id;
-
-            read(p, buf, 4);
-            memcpy(&signal_id, buf, 4);
-            
-            printf("[P2] Otrzymalem SIGUSR1 oraz ID: %d\n", signal_id);
-            switch(signal_id) {
-                case SIGTSTP:
-                    printf("[P2] Zatrzymano proces (SIGTSTP). Oczekiwanie na SIGCONT...\n");
-                    is_paused = 1;
-                    break;
-                case SIGINT:
-                    printf("[P2] Wznowiono proces (SIGCONT)\n");
-                    is_paused = 0;
-                    break;
-                case SIGTERM:
-                break;
-            }
-        break;
-        default:
-            kill(getppid(), s);
-        break;
-    }
     got_signal = 0;
+    if (s == SIGUSR1) {
+        int signal_id;
+        if (read(p, &signal_id, sizeof(int)) == sizeof(int)) {
+            printf("[P2] Przetwarzam ID sygnalu: %d\n", signal_id);
+            if (signal_id == SIGTSTP) {
+                printf("[P2] Pauza\n");
+                is_paused = 1;
+            } else if (signal_id == SIGINT || signal_id == SIGCONT) {
+                printf("[P2] Wznowienie\n");
+                is_paused = 0;
+            }
+            // Propagacja
+            kill(p1_id, SIGUSR1);
+        }
+    } else {
+        // Zewnętrzny - do maina
+        printf("[P2] Forwarding signal %d to parent\n", s);
+        kill(getppid(), s);
+    }
 }
